Other: seperate the tests
